Move FileMimeTypesModule to play-java and add it to reference.conf instead of BuiltinModule
