+++ framework/src/play-java/src/main/resources/reference.conf
@@ -1,6 +1,7 @@
 play {
   modules {
     enabled += "play.inject.BuiltInModule"
+    enabled += "play.core.FileMimeTypesModule"
     enabled += "play.core.ObjectMapperModule"
     enabled += "play.routing.RoutingDslModule"
   }
