+++ framework/src/play/src/main/java/play/mvc/FileMimeTypes.java
@@ -6,6 +6,7 @@ package play.mvc;
 
 import scala.compat.java8.OptionConverters;
 
+import javax.inject.Inject;
 import javax.inject.Singleton;
 import java.util.Optional;
 
@@ -14,6 +15,7 @@ public class FileMimeTypes {
 
     private final play.api.http.FileMimeTypes fileMimeTypes;
 
+    @Inject
     public FileMimeTypes(play.api.http.FileMimeTypes fileMimeTypes) {
         this.fileMimeTypes = fileMimeTypes;
     }
